requirements.txt
flask
flask-smorest
python-dotenv
Flask-SQLAlchemy
marshmallow
flask-jwt-extended
passlib
flask-migrate
gunicorn>=21.2.0
psycopg[binary]
requests
flask-cors